Make logic MetaTensor only. Update Inference
- Inside the modeling we now purely work with MetaTensore, instead of
  numpy arrays. numpy arrays dont include meta information (e.g. affine)
and "guessing" this meta information would result in hard-to-decect long
time errors. If a user has only numpy arrays it is on them to think
about which affine information they should add, and the risks involved
with that.

- the _prepare_data_point method is the single entry point to the
  modules. It works now independently for scans and masks, so we could
for example pass a Path to the scan and a Meta tensor for the
segmentation.

- the feature extractor does not sort lesions based on their volume
  anymore.

- I removed all world coordinates (very confusing!). Instead I assign
  the a custom lesion id before feature extraction, which will be
propagated up to prediction and can be remapped later.